GP-1083: Set parent activity for TM and postal return imports
This sets the custom activity field "parent_activity_id" for any
activity created by the TM or postal return importer.

For TM, the parent activity is determined by looking for the most recent
activity with the same campaign of the activity type "Action" and the
activity date somewhere between the date of the call and 90 days prior.

For postal returns, the parent activity is the most recent activity
of the same campaign with activity medium "Letter Mail".