Add PostAT return handler
This adds a new PostAT postal return handler. This handler is
fairly similar to the existing PostRetourRecordHandler, so this
includes a refactoring to extract much of the code into a new
base class that is used by both postal return handlers.

Some of the existing business logic was changed: We now use the
activity date of postal mailings to determine whether an address
has been changed, unless the mailing cannot be found.