GP-47846 Fix contract activity parent update
This fixes an issue where parent activities of contract activities
are sometimes set to a TM activity even if the contract activity
was not triggered by the import/TM activity at all.

Previously, the import attempted to find related contract activities
by simply looking up the latest contract activity for a membership.
This produces false positives e.g. if a membership has already
been cancelled through other means, causing that previous cancellation
to be linked to the TM activity.

This changes the import so that we set the parent field only for the
exact activity returned by a CE API call triggered by the import,
except in the case of new memberships (where there cannot be
any existing unrelated contract activities).